# perf: experiment with native (Rust) CRC16 implementation