--- a/packages/zwave-js/src/lib/controller/Controller.ts
+++ b/packages/zwave-js/src/lib/controller/Controller.ts
@@ -251,7 +251,7 @@ import { ZWaveLibraryTypes } from "./ZWaveLibraryTypes";
 import { protocolVersionToSDKVersion } from "./ZWaveSDKVersions";
 
 export type HealNodeStatus = "pending" | "done" | "failed" | "skipped";
-export type SerialAPIVersion =
+export type SDKVersion =
 	| `${number}.${number}`
 	| `${number}.${number}.${number}`;
 
@@ -321,14 +321,14 @@ export class ZWaveController extends TypedEventEmitter<ControllerEventCallbacks>
 		);
 	}
 
-	private _libraryVersion: string | undefined;
-	public get libraryVersion(): string | undefined {
-		return this._libraryVersion;
-	}
-
 	private _type: ZWaveLibraryTypes | undefined;
 	public get type(): ZWaveLibraryTypes | undefined {
 		return this._type;
+	}
+
+	private _sdkVersion: string | undefined;
+	public get sdkVersion(): string | undefined {
+		return this._sdkVersion;
 	}
 
 	private _homeId: number | undefined;
@@ -373,48 +373,39 @@ export class ZWaveController extends TypedEventEmitter<ControllerEventCallbacks>
 		return this._isSlave;
 	}
 
-	private _serialApiVersion: string | undefined;
-	public get serialApiVersion(): string | undefined {
-		return this._serialApiVersion;
-	}
-
-	/** Checks if the Serial API version is greater than the given one */
-	public serialApiGt(version: SerialAPIVersion): boolean | undefined {
-		// TODO: Rename these to sdkVersionGt(e) etc...
-		if (this._libraryVersion === undefined) {
+	/** Checks if the SDK version is greater than the given one */
+	public sdkVersionGt(version: SDKVersion): boolean | undefined {
+		if (this._sdkVersion === undefined) {
 			return undefined;
 		}
-		const sdkVersion = protocolVersionToSDKVersion(this._libraryVersion);
+		const sdkVersion = protocolVersionToSDKVersion(this._sdkVersion);
 		return semver.gt(padVersion(sdkVersion), padVersion(version));
 	}
 
-	/** Checks if the Serial API version is greater than or equal to the given one */
-	public serialApiGte(version: SerialAPIVersion): boolean | undefined {
-		// TODO: Rename these to sdkVersionGt(e) etc...
-		if (this._libraryVersion === undefined) {
+	/** Checks if the SDK version is greater than or equal to the given one */
+	public sdkVersionGte(version: SDKVersion): boolean | undefined {
+		if (this._sdkVersion === undefined) {
 			return undefined;
 		}
-		const sdkVersion = protocolVersionToSDKVersion(this._libraryVersion);
+		const sdkVersion = protocolVersionToSDKVersion(this._sdkVersion);
 		return semver.gte(padVersion(sdkVersion), padVersion(version));
 	}
 
-	/** Checks if the Serial API version is lower than the given one */
-	public serialApiLt(version: SerialAPIVersion): boolean | undefined {
-		// TODO: Rename these to sdkVersionGt(e) etc...
-		if (this._libraryVersion === undefined) {
+	/** Checks if the SDK version is lower than the given one */
+	public sdkVersionLt(version: SDKVersion): boolean | undefined {
+		if (this._sdkVersion === undefined) {
 			return undefined;
 		}
-		const sdkVersion = protocolVersionToSDKVersion(this._libraryVersion);
+		const sdkVersion = protocolVersionToSDKVersion(this._sdkVersion);
 		return semver.lt(padVersion(sdkVersion), padVersion(version));
 	}
 
-	/** Checks if the Serial API version is lower than or equal to the given one */
-	public serialApiLte(version: SerialAPIVersion): boolean | undefined {
-		// TODO: Rename these to sdkVersionGt(e) etc...
-		if (this._libraryVersion === undefined) {
+	/** Checks if the SDK version is lower than or equal to the given one */
+	public sdkVersionLte(version: SDKVersion): boolean | undefined {
+		if (this._sdkVersion === undefined) {
 			return undefined;
 		}
-		const sdkVersion = protocolVersionToSDKVersion(this._libraryVersion);
+		const sdkVersion = protocolVersionToSDKVersion(this._sdkVersion);
 		return semver.lte(padVersion(sdkVersion), padVersion(version));
 	}
 
@@ -431,6 +422,11 @@ export class ZWaveController extends TypedEventEmitter<ControllerEventCallbacks>
 	private _productId: number | undefined;
 	public get productId(): number | undefined {
 		return this._productId;
+	}
+
+	private _firmwareVersion: string | undefined;
+	public get firmwareVersion(): string | undefined {
+		return this._firmwareVersion;
 	}
 
 	private _supportedFunctionTypes: FunctionType[] | undefined;
@@ -478,7 +474,7 @@ export class ZWaveController extends TypedEventEmitter<ControllerEventCallbacks>
 	public supportsFeature(feature: ZWaveFeature): boolean | undefined {
 		switch (feature) {
 			case ZWaveFeature.SmartStart:
-				return this.serialApiGte(minFeatureVersions[feature]);
+				return this.sdkVersionGte(minFeatureVersions[feature]);
 		}
 	}
 
@@ -695,14 +691,14 @@ export class ZWaveController extends TypedEventEmitter<ControllerEventCallbacks>
 					supportCheck: false,
 				},
 			);
-		this._serialApiVersion = apiCaps.serialApiVersion;
+		this._firmwareVersion = apiCaps.firmwareVersion;
 		this._manufacturerId = apiCaps.manufacturerId;
 		this._productType = apiCaps.productType;
 		this._productId = apiCaps.productId;
 		this._supportedFunctionTypes = apiCaps.supportedFunctionTypes;
 		this.driver.controllerLog.print(
 			`received API capabilities:
-  serial API version:  ${this._serialApiVersion}
+  firmware version:    ${this._firmwareVersion}
   manufacturer ID:     ${num2hex(this._manufacturerId)}
   product type:        ${num2hex(this._productType)}
   product ID:          ${num2hex(this._productId)}
@@ -729,12 +725,12 @@ export class ZWaveController extends TypedEventEmitter<ControllerEventCallbacks>
 					supportCheck: false,
 				},
 			);
-		this._libraryVersion = version.libraryVersion;
+		this._sdkVersion = version.sdkVersion;
 		this._type = version.controllerType;
 		this.driver.controllerLog.print(
 			`received version info:
   controller type: ${ZWaveLibraryTypes[this._type]}
-  library version: ${this._libraryVersion}`,
+  library version: ${this._sdkVersion}`,
 		);
 
 		this.driver.controllerLog.print(
@@ -948,7 +944,7 @@ export class ZWaveController extends TypedEventEmitter<ControllerEventCallbacks>
 			getFirmwareVersionsMetadata(),
 		);
 		controllerValueDB.setValue(getFirmwareVersionsValueId(), [
-			this._serialApiVersion,
+			this._firmwareVersion,
 		]);
 
 		if (
@@ -4536,7 +4532,7 @@ ${associatedNodes.join(", ")}`,
 
 		let ret: Buffer;
 		try {
-			if (this.serialApiGte("7.0")) {
+			if (this.sdkVersionGte("7.0")) {
 				ret = await this.backupNVMRaw700(onProgress);
 			} else {
 				ret = await this.backupNVMRaw500(onProgress);
@@ -4665,7 +4661,7 @@ ${associatedNodes.join(", ")}`,
 				"Converting NVM to target format...",
 			);
 			let targetNVM: Buffer;
-			if (this.serialApiGte("7.0")) {
+			if (this.sdkVersionGte("7.0")) {
 				targetNVM = await this.backupNVMRaw700(convertProgress);
 			} else {
 				targetNVM = await this.backupNVMRaw500(convertProgress);
@@ -4673,7 +4669,7 @@ ${associatedNodes.join(", ")}`,
 			const convertedNVM = migrateNVM(nvmData, targetNVM);
 
 			this.driver.controllerLog.print("Restoring NVM backup...");
-			if (this.serialApiGte("7.0")) {
+			if (this.sdkVersionGte("7.0")) {
 				await this.restoreNVMRaw700(convertedNVM, restoreProgress);
 			} else {
 				await this.restoreNVMRaw500(convertedNVM, restoreProgress);
@@ -4725,7 +4721,7 @@ ${associatedNodes.join(", ")}`,
 		}
 
 		try {
-			if (this.serialApiGte("7.0")) {
+			if (this.sdkVersionGte("7.0")) {
 				await this.restoreNVMRaw700(nvmData, onProgress);
 			} else {
 				await this.restoreNVMRaw500(nvmData, onProgress);