--- a/packages/zwave-js/src/lib/driver/CommandQueueMachine.ts
+++ b/packages/zwave-js/src/lib/driver/CommandQueueMachine.ts
@@ -330,8 +330,6 @@ export function createCommandQueueMachine(
 		{
 			services: {
 				executeSerialAPICommand: (ctx) => {
-					// If there is an error while creating the command machine (e.g. during message serialization)
-					// wrap it in a rejected promise, so xstate can handle it
 					try {
 						return createSerialAPICommandMachine(
 							ctx.currentTransaction!.parts.current!,
@@ -339,6 +337,8 @@ export function createCommandQueueMachine(
 							params,
 						);
 					} catch (e) {
+						// If there is an error while creating the command machine (e.g. during message serialization)
+						// wrap it in a rejected promise, so xstate can handle it
 						implementations.log(
 							`Unexpected error during SerialAPI command: ${getErrorMessage(
 								e,