… docs/_sidebar.md
… docs/api/CCs/_sidebar.md
… docs/api/controller.md
… docs/api/driver.md
… docs/api/node.md
… docs/getting-started/migrating-to-v9.md
… packages/core/package.json
… packages/core/src/util/crc.test.ts
… packages/zwave-js/src/lib/node/Node.ts
… packages/zwave-js/src/lib/node/NodeStatistics.ts
… yarn.lock