CRC16 isn't used a ton, but even with only 10 bytes payload, the Rust implementation is ~40% faster, scaling up to 32x for larger payloads:
```
PERF JS:
10 bytes: 0.05 ms = 204.30 KB/s, result = 0xbab9
100 bytes: 0.06 ms = 1.60 MB/s, result = 0xc427
1000 bytes: 1.99 ms = 490.88 KB/s, result = 0x829e
10000 bytes: 1.64 ms = 5.81 MB/s, result = 0x8c17
100000 bytes: 6.27 ms = 15.21 MB/s, result = 0x7b4
1000000 bytes: 64.61 ms = 14.76 MB/s, result = 0xe416

PERF RUST:
10 bytes: 0.03 ms = 292.38 KB/s, result = 0xbab9
100 bytes: 0.01 ms = 9.26 MB/s, result = 0xc427
1000 bytes: 0.00 ms = 207.32 MB/s, result = 0x829e
10000 bytes: 0.02 ms = 431.53 MB/s, result = 0x8c17
100000 bytes: 0.20 ms = 474.94 MB/s, result = 0x7b4
1000000 bytes: 2.02 ms = 471.72 MB/s, result = 0xe416
```

This can serve as a nice experiment to investigate where it makes sense to defer CPU intensive tasks to Rust.