@AlCalzone 2022-02-08
Not sure this is worth the additional complexity